Fix generate:readme gulp task
When we used vinylPaths and js-yaml together in a pipe, it started
breaking with "UnhandledPromiseRejectionWarning: Unhandled promise
rejection (rejection id: 1): Error: write callback called multiple
times"

As we're already using gulp-data we can use it to pipe data from yaml to
nunjucks template, as gulp-data exposes file paths

Generate:readme task also requires capitaliseComponentName function as
an environment filter, which is used in the nunjucks template macro.

capitaliseComponentName is used both the the gulp task and app.js
to it makes sense to move it out to /lib as a helper function and
then import it where it is needed.
